+++ src/lib.rs
@@ -4,18 +4,25 @@ extern crate rocket;
 pub mod models;
 pub mod utils {
     pub mod account;
+    pub mod asset;
+    pub mod category;
+    pub mod contest;
     pub mod organization;
     pub mod problem;
     pub mod session;
-    pub mod category;
+    pub mod submission;
 }
 
 pub mod routes {
     pub mod account;
+    pub mod asset;
     pub mod category;
+    pub mod contest;
     pub mod index;
     pub mod organization;
+
     pub mod problem;
+    pub mod submission;
 }
 
 pub mod cors;