+++ src/models/mod.rs
@@ -1,9 +1,12 @@
 pub mod account;
+pub mod asset;
+pub mod category;
+pub mod contest;
 pub mod error;
 pub mod organization;
 pub mod problem;
 pub mod response;
 pub mod shared;
-pub mod category;
+pub mod submission;
 
 pub use shared::*;